Updated point size and line width to floats (see OpenGL spec) Use our own BlockList within DataNode (efficiency + we can control things better)

+++ core/src/org/vast/stt/style/LinePointGraphic.java
@@ -30,7 +30,7 @@ package org.vast.stt.style;
  */
 public class LinePointGraphic extends PrimitiveGraphic
 {
-    public int width = 1;
+    public float width = 1.0f;
     public boolean lineBreak;
     public boolean smooth = false;
 }
